Actually support the C++ stdlib
Now that we're downloading arbitrary versions of clang, if we want to
actually use the C++ standard library in cross-compilation
environments, we need to compile it too.

bazel doesn't make that super easy for now, as there is no facility
for transitioning between a no-stdlib-C++ compiler and a
with-stdlib-C++ compiler automatically.  Thus we currently require
cc_binary nodes to manually link in the libcxx label.